jobs/job_template/requirements.txt
python-dotenv>=1.0.0
requests>=2.0
pytest>=7.0
